Have jinja2 and handlebars use the same template_helper_from_function method.
