+++ python/semantic_kernel/prompt_template/utils/__init__.py
@@ -1,13 +1,11 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-from semantic_kernel.prompt_template.utils.handlebars_function_helpers import create_handlebars_helper_from_function
 from semantic_kernel.prompt_template.utils.handlebars_system_helpers import HANDLEBAR_SYSTEM_HELPERS
-from semantic_kernel.prompt_template.utils.jinja2_function_helpers import create_jinja2_helper_from_function
 from semantic_kernel.prompt_template.utils.jinja2_system_helpers import JINJA2_SYSTEM_HELPERS
+from semantic_kernel.prompt_template.utils.template_function_helpers import create_template_helper_from_function
 
 __all__ = [
-    "create_handlebars_helper_from_function",
-    "create_jinja2_helper_from_function",
+    "create_template_helper_from_function",
     "HANDLEBAR_SYSTEM_HELPERS",
     "JINJA2_SYSTEM_HELPERS",
 ]
